requirements.txt
plotly-express
streamlit
pandas